requirements.txt
asttokens==2.0.5
attrs==19.3.0
Automat==0.8.0
backcall==0.2.0
bcrypt==3.2.0
black==21.12b0
blinker==1.4
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
cliapp==1.20180812.1
click==8.0.3
cloud-init==21.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
decorator==5.1.1
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
executing==0.8.2
Flask==1.0
Flask-Assets==2.0
Flask-Bcrypt==1.0.1
Flask-DebugToolbar==0.11.0
Flask-SQLAlchemy==2.2
Flask-WTF==1.0.1
forex-python==1.8
greenlet==1.1.2
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
ipython==8.0.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
Markdown==3.1.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
more-itertools==4.2.0
mypy-extensions==0.4.3
netifaces==0.10.4
neuralart==1.1.0
numpy==1.22.0
oauthlib==3.1.0
packaging==20.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.4.1
prompt-toolkit==3.0.24
psycopg2-binary==2.9.3
pure-eval==0.2.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.21
Pygments==2.11.2
PyGObject==3.36.0
PyHamcrest==1.9.0
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.5
python-debian===0.1.36ubuntu1
PyYAML==5.3.1
requests==2.27.1
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.16.0
sos==4.1
spotipy==2.19.0
SQLAlchemy==1.4.31
ssh-import-id==5.10
stack-data==0.1.3
systemd-python==234
tomli==1.2.3
torch==1.10.1
traitlets==5.1.1
ttystatus==0.38
Twisted==18.9.0
typing-extensions==4.0.1
ubuntu-advantage-tools==27.2
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.9
wadllib==1.3.3
wcwidth==0.2.5
webassets==2.0
Werkzeug==2.0.2
WTForms==3.0.1
zipp==1.0.0
zope.interface==4.7.1
